Refactor: 칸반 리스트 컴포넌트 리팩토링 #39
